Add readme and copy of GPLv2
Move the upstream readme aside, so we can show a sensible readme to a
user of the driver.

The readme explains where I got the code from, what I've done to it,
and how to use the module.

The existing files contain headers indicating they are GPLv2 or later.
In actual fact, they appear to be a copy of cdc-acm.c from the 3.6
kernel with the original copyrights and comments stripped.

Signed-off-by: Joel Stanley <joel@jms.id.au>